# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.ChangePasswordRequest:
    properties:
      newPassword:
        maxLength: 64
        minLength: 6
        type: string
      oldPassword:
        maxLength: 64
        minLength: 6
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDictDataRequest:
    properties:
      dictTypeId:
        description: |-
          DictTypeID 决定归属；dict_type 冗余列由服务层按此 ID 推导，不由前端传入，
          避免两者不一致。
        type: integer
      isDefault:
        enum:
        - 0
        - 1
        type: integer
      label:
        maxLength: 100
        type: string
      remark:
        maxLength: 255
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      value:
        maxLength: 100
        type: string
    required:
    - dictTypeId
    - label
    - value
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDictTypeRequest:
    properties:
      name:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        description: |-
          Type 是数据项的归属键，只允许字母/数字/下划线：
          它会作为 URL 路径段用于按类型取数据（/dicts/data/{type}）。
        maxLength: 64
        type: string
    required:
    - name
    - type
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.LoginRequest:
    properties:
      captchaCode:
        type: string
      captchaId:
        description: CaptchaID 与 CaptchaCode 配对校验；未启用验证码时可为空。
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      username:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDictDataRequest:
    properties:
      isDefault:
        enum:
        - 0
        - 1
        type: integer
      label:
        maxLength: 100
        type: string
      remark:
        maxLength: 255
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      value:
        maxLength: 100
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDictTypeRequest:
    properties:
      name:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        maxLength: 64
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.DictDataItem:
    properties:
      createdAt:
        type: string
      dictType:
        type: string
      dictTypeId:
        type: integer
      id:
        type: integer
      isDefault:
        type: integer
      label:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
      value:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.DictTypeItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      status:
        type: integer
      type:
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.LoginResult:
    properties:
      accessToken:
        type: string
      expiresIn:
        description: ExpiresIn 为 Access Token 剩余有效秒数。
        type: integer
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 基于 RBAC 的通用后台管理系统接口
  title: 权限管理后台 API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.LoginResult'
              type: object
      summary: 用户登录
      tags:
      - 认证
...
  /auth/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 密码参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改自己的密码
      tags:
      - 个人中心
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: 刷新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.LoginResult'
              type: object
      summary: 刷新令牌
      tags:
      - 认证
...
  /dicts/data:
    get:
      parameters:
      - description: DictType 为空时返回全部类型的数据；前端双栏布局下通常按类型过滤。
        in: query
        maxLength: 64
        name: dictType
        type: string
      - in: query
        maxLength: 100
        name: label
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.DictDataItem'
                        type: array
                    type: object
              type: object
      summary: 字典数据列表
      tags:
      - 字典管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 字典数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDictDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增字典数据
      tags:
      - 字典管理
  /dicts/data/{id}:
    delete:
      parameters:
      - description: 字典数据ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除字典数据
      tags:
      - 字典管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 字典数据ID
        in: path
        name: id
        required: true
        type: integer
      - description: 字典数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDictDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改字典数据
      tags:
      - 字典管理
  /dicts/data/type/{type}:
    get:
      parameters:
      - description: 字典类型标识
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.DictDataItem'
                  type: array
              type: object
      summary: 按类型取字典项
      tags:
      - 字典管理
  /dicts/types:
    get:
      parameters:
      - in: query
        maxLength: 64
        name: name
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - in: query
        maxLength: 64
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.DictTypeItem'
                        type: array
                    type: object
              type: object
      summary: 字典类型列表
      tags:
      - 字典管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 字典类型
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateDictTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增字典类型
      tags:
      - 字典管理
  /dicts/types/{id}:
    delete:
      parameters:
      - description: 字典类型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除字典类型
      tags:
      - 字典管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 字典类型ID
        in: path
        name: id
        required: true
        type: integer
      - description: 字典类型
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateDictTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改字典类型
      tags:
      - 字典管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"